Added JWT authentication and user-specific task management

+++ controllers/taskController.js
@@ -1,40 +1,46 @@
 const con=require('../db/db');
 
-const tasks=async(req,res)=>{
-   const page=parseInt(req.query.page) || 1;
-   const limit=parseInt(req.query.limit) || 5;
-   const status=req.query.status;
-   const offset=(page-1)*limit;
-   let result;
-   let result1;
-   console.log(page, limit , offset);
-   try{
-      if(status){
-          result=await con.query("select * from task where status=$1 limit $2 offset $3",[status,limit,offset]);
-          result1=await con.query("select count(*) from task where status=$1",[status]);
-      }else{
-          result=await con.query('select * from task limit $1 offset $2',[limit,offset]);
-          result1=await con.query("select count(*) from task");
+
+ const tasks=async(req,res)=>{
+   
+      const page=parseInt(req.query.page) || 1;
+      const limit=parseInt(req.query.limit) || 5;
+      const status=req.query.status;
+      const offset=(page-1)*limit;
+      const user_id=req.user.id;
+      let result;
+      let result1;
+      console.log(page, limit , offset);
+      try{
+         if(status){
+            result=await con.query("select * from task where status=$1 and user_id=$2 limit $3 offset $4",[status,user_id,limit,offset]);
+            result1=await con.query("select count(*) from task where status=$1 and user_id=$2",[status,user_id]);
+         }else{
+            result=await con.query('select * from task where user_id=$1 limit $2 offset $3 ',[user_id,limit,offset]);
+            result1=await con.query("select count(*) from task where user_id=$1 ",[user_id]);
+         }
+         let totalTasks=parseInt(result1.rows[0].count);
+         let totalPages=Math.ceil(totalTasks/limit);
+         res.send({success:true,
+            page:page
+            ,limit:limit,
+            totalTasks:totalTasks,
+            totalPages:totalPages,
+            data:result.rows});
+         
+      }catch(e){
+      console.log(e.message);
+      res.status(500).send("Server Error")
       }
-      let totalTasks=parseInt(result1.rows[0].count);
-      let totalPages=Math.ceil(totalTasks/limit);
-      res.send({success:true,
-         page:page
-         ,limit:limit,
-         totalTasks:totalTasks,
-         totalPages:totalPages,
-         data:result.rows});
+      console.log(user_id);
       
-   }catch(e){
-    console.log(e.message);
-    res.status(500).send("Server Error")
-   }
-};
+   };
 
 const addtasks=async(req,res)=>{
     const{title,description}=req.body;
+      const user_id=req.user.id;
     try{
-     const result=await con.query("insert into task(title,description)values($1,$2) returning *",[title,description])
+     const result=await con.query("insert into task(title,description,user_id)values($1,$2,$3) returning *",[title,description,user_id])
      if(result.rowCount>0){
         res.send({msg:'Task Inserted'});
      }else{
@@ -45,13 +51,16 @@ const addtasks=async(req,res)=>{
         console.log(e.message);
         res.status(500).send("Server Error")
     }
+    console.log(user_id);
+    
 };
 
 const updttask=async(req,res)=>{
     const id=parseInt(req.params.id);
     const{title,description,status}=req.body;
+    const user_id=req.user.id;
     try{
-       const result=await con.query("update task set title=$1,description=$2,status=$3,updated_at=CURRENT_TIMESTAMP where id=$4 returning *",[title,description,status,id]);
+       const result=await con.query("update task set title=$1,description=$2,status=$3,updated_at=CURRENT_TIMESTAMP where id=$4 and user_id=$5 returning *",[title,description,status,id,user_id]);
        if(result.rowCount>0){
         res.send({msg:"Task Updated",data:result.rows[0]})
        }else{
@@ -66,11 +75,12 @@ const updttask=async(req,res)=>{
 
 const deletetask=async(req,res)=>{
       const id=parseInt(req.params.id);
+      const user_id=req.user.id;
       if (isNaN(id)) {
   return res.status(400).json({ msg: "Invalid ID" });
 }
       try{
-         const result=await con.query("delete from task where id=$1 returning *",[id]);
+         const result=await con.query("delete from task where id=$1 and user_id=$2 returning *",[id,user_id]);
          if(result.rowCount>0){
             res.send({msg:"Task Deleted",data:result.rows[0]});
          }else{
@@ -84,13 +94,14 @@ const deletetask=async(req,res)=>{
 };
 
 const filtertask=async(req,res)=>{ 
-   const status=req.query.status; 
+   const status=req.query.status;
+   const user_id=req.user.id; 
    let result; 
    try{ 
       if(status){ 
          result=await con.query("Select * from task where status=$1 ",[status]);
        }else{
-          result=await con.query("Select * from task"); 
+          result=await con.query("Select * from task where user_id=$1",[user_id]); 
 
        } 
        res.send({success:true ,data:result.rows})
